refactor: Change the folder structure
